add validation for prize input

+++ app/controllers/prizes_controller.rb
@@ -20,13 +20,18 @@ class PrizesController < ApplicationController
     end
 
     post '/prizes' do
+        @user = current_user
         #do I need login validation for post routes??
-        @horse = Horse.find(params[:horse_id])
-        @horseshow = Horseshow.find_or_create_by(params[:horseshow])
+        if prize_valid?(params)
+            @horse = Horse.find(params[:horse_id])
+            @horseshow = Horseshow.find_or_create_by(params[:horseshow])
         
-        @prize = Prize.create(point_total: params[:point_total], horse_id: @horse.id, horseshow_id: @horseshow.id, user_id: @horse.user_id)
+            @prize = Prize.create(point_total: params[:point_total], horse_id: @horse.id, horseshow_id: @horseshow.id, user_id: @horse.user_id)
         
-        redirect "/account"
+            redirect "/users/#{@user.slug}"
+        else
+            "Prize input is invalid" #link to an error page??
+        end
     end
 
     get '/prizes/:id' do
@@ -53,17 +58,26 @@ class PrizesController < ApplicationController
     patch '/prizes/:id' do
         @user = current_user
         @prize = Prize.find(params[:id])
-        @horseshow = Horseshow.find_or_create_by(params[:horseshow])
-        @horse = Horse.find(params[:horse_id])
-        @prize.update!(point_total: params[:point_total], horseshow_id: @horseshow.id, horse_id: @horse.id, user_id: @user.id)
 
-        redirect "/prizes"
+        if prize_valid?(params)
+            @horseshow = Horseshow.find_or_create_by(params[:horseshow])
+            @horse = Horse.find(params[:horse_id])
+            @prize.update(point_total: params[:point_total], horseshow_id: @horseshow.id, horse_id: @horse.id, user_id: @user.id)
+
+            redirect "/prizes"
+        else
+            "Prize input is invalid"
+        end
     end
 
     delete '/prizes/:id' do
-        prize = Prize.find(params[:id])
-        prize.destroy
+        if logged_in?
+            prize = Prize.find(params[:id])
+            prize.destroy
 
-        redirect to "/prizes"
+            redirect to "/prizes"
+        else
+            redirect "/login"
+        end
     end
 end
